Persist rules for non-existing files
The behaviour of the library has been changed to add rules for
non-existing files if strict path check is false.
Now, the file audit rule for a non-existing file will be added
successfully if its immediate parent directory exists. It will
still fail for non-existing files if the immediate parent does
not exist.
This behaviour is consistent with the `auditctl`